+++ NEWS.rst
@@ -5,6 +5,7 @@ NEWS for system-image updater
 0.7 (2013-XX-XX)
 ================
  * No reboot should be issued if there is no update available.  (LP: #1202915)
+ * DBus API implemented.  (LP: #1192585)
 
 0.6 (2013-07-15)
 ================